style: Проставили документирующие комментарии refactor: Вывели валидацию фильмов и пользователей в отдельный класс
