--- src/main/java/ru/yandex/practicum/filmorate/Exception/InvalidEmailException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.Exception;
-
-public class InvalidEmailException extends RuntimeException {
-    public InvalidEmailException(String s) {
-        super(s);
-    }
-}
