--- src/main/java/ru/yandex/practicum/filmorate/Exception/UserAlreadyExistException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.Exception;
-
-public class UserAlreadyExistException extends RuntimeException {
-    public UserAlreadyExistException(String s) {
-        super(s);
-    }
-}
